Refactored out chart part of t_figure
